fixed warning
git-svn-id: http://svn.xp-dev.com/svn/wxextension@3677 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/src/stc.cpp
@@ -645,7 +645,7 @@ const wxString wxExSTC::GetFindString(int search_flags) const
 
     if (flags & wxSTC_FIND_REGEXP)
     {
-      for (int i = 0; i < selection.size() && alnum; i++)
+      for (size_t i = 0; i < selection.size() && alnum; i++)
       {
         if (!isalnum(selection[i]))
         {
